moved the java script code out from the vue components
