+++ src/components/auth/login/LoginPage.vue
@@ -1,5 +1,5 @@
 <style lang="css" scoped src="./login.css"></style>
-
+<script src="./loginPage.js"></script>
 <template>
   <div class="wrapper">
     <header>Login Form</header>
@@ -28,65 +28,3 @@
     <div class="sign-txt">Not yet a member? <router-link to="/signIn">Signup now</router-link></div>
   </div>
 </template>
-
-<script>
-import axios from 'axios'
-
-export default {
-    name:"LoginPage",
-  data() {
-    return {
-      email: '',
-      password: '',
-      emailError: false,
-      passwordError: false
-    };
-  },
-  computed: {
-    validEmail() {
-      const pattern = /^[^ ]+@[^ ]+\.[a-z]{2,3}$/;
-      return this.email.match(pattern);
-    }
-  },
-  methods: {
-    submitForm() {
-      this.emailError = this.email === '';
-      this.passwordError = this.password === '';
-
-      if (!this.emailError && !this.passwordError) {
-       this.logIn();
-      }
-    },
-        initUserDatas(){
-          const body={
-    "id": 1,
-    "name": "nandyRps",
-    "email": "prasannaNandhini730@gmail.com",
-    "password": "prasaNan57",
-    "userProfile": {
-        "userProfile": 2,
-        "userName": "nandyRps",
-        "status": "beginner"
-    }
-}
-         localStorage.setItem("user-info",JSON.stringify(body));
-          
-         },
-   async logIn(){
-    console.log(this.email,this.password);
-    try {   const response=  await axios.get(`http://localhost:8090/login/email=${this.email}&password=${this.password}`);
-       if(response.data!=''){
-          console.log(response.data);
-           localStorage.setItem('user-info',JSON.stringify(response.data));
-            this.$router.push({ name: 'HomePage' })
-           alert("Login Success");
-        } else {alert("InValid Username or Password")}
-    } catch(err){
-          console.log(err);
-          this.initUserDatas(); 
-          this.$router.push({ name: 'HomePage' })
-    }
-   }
-  }
-};
-</script>
