+++ src/components/demo/demo.css
@@ -1,5 +1,3 @@
-
-
 .card-body{
    border-radius: 10px !important;
 }
