+++ src/components/footer/footerPage.js
@@ -1,0 +1,3 @@
+export default {
+    name:'footerPage'
+}
